+++ tests/Feature/Auth/RegistrationTest.php
@@ -34,6 +34,6 @@ class RegistrationTest extends TestCase
         ]);
 
         $this->assertAuthenticated();
-        $response->assertRedirect(route('evolayer.base.home', absolute: false));
+        $response->assertRedirect(route('home', absolute: false));
     }
 }