+++ tests/Feature/ExampleTest.php
@@ -4,8 +4,8 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 
 uses(RefreshDatabase::class);
 
-test('home page returns a successful response', function () {
-    $this->get(route('home'))
+test('public landing page returns a successful response', function () {
+    $this->get(route('welcome'))
         ->assertOk()
-        ->assertSee('"component":"evolayer\/about"', false);
+        ->assertSee('"component":"evolayer\/base"', false);
 });