+++ tests/Feature/Settings/ProfileUpdateTest.php
@@ -73,7 +73,7 @@ class ProfileUpdateTest extends TestCase
 
         $response
             ->assertSessionHasNoErrors()
-            ->assertRedirect(route('home'));
+            ->assertRedirect(route('welcome'));
 
         $this->assertGuest();
         $this->assertNull($user->fresh());